# Conflicts:
#	dt-core/setup-functions.php
#	functions.php

+++ dt-mapping/migrations/0015-v2-migrate-custom-locations.php
@@ -3,6 +3,7 @@ if ( ! defined( 'ABSPATH' ) ) { exit; } // Exit if accessed directly
 
 /**
  * Class DT_Mapping_Module_Migration_0015
+ * 2nd step: copy custom data back to location grid table
  *
  * @version_added 1.30.2
  */
@@ -13,11 +14,15 @@ class DT_Mapping_Module_Migration_0015 extends DT_Mapping_Module_Migration
      * @throws \Exception  Got error when creating table $name.
      */
     public function up() {
+        if ( version_compare( dt_get_initial_install_meta( 'theme_version' ), '1.31.0', '>=' ) ){
+            return;
+        }
+
         global $wpdb;
         $has_upgrade_table = $wpdb->get_var( "SHOW TABLES LIKE '{$wpdb->prefix}dt_location_grid_upgrade';" );
         $location_grid_rows = (int) $wpdb->get_var( "SELECT count(*) FROM $wpdb->dt_location_grid" );
         if ( !empty( $has_upgrade_table ) && !empty( $location_grid_rows ) ) {
-            $wpdb->query( "INSERT INTO `{$wpdb->prefix}dt_location_grid` SELECT * FROM `{$wpdb->prefix}dt_location_grid_upgrade` WHERE grid_id > 1000000000;" );
+            $wpdb->query( "INSERT INTO $wpdb->dt_location_grid SELECT * FROM `{$wpdb->prefix}dt_location_grid_upgrade` WHERE grid_id > 1000000000;" );
             $wpdb->query( "DROP TABLE IF EXISTS `{$wpdb->prefix}dt_location_grid_upgrade`;" );
         }
     }
@@ -26,7 +31,6 @@ class DT_Mapping_Module_Migration_0015 extends DT_Mapping_Module_Migration
      * @throws \Exception  Got error when dropping table $name.
      */
     public function down() {
-
     }
 
     /**
@@ -42,5 +46,4 @@ class DT_Mapping_Module_Migration_0015 extends DT_Mapping_Module_Migration
      */
     public function test() {
     }
-
 }
